Create MACHINE PROBLEM #6.1 - Germain Primes.c
Ask the user to enter a number, say n.  

Determine if this number is a Germain Prime or not.  A number n is said to be Germain if it is prime and 2*n + 1 is also prime.  

Recall that you have solved this through a flowchart already. This time, transform your solution from flowchart to C Programming, strictly using the while-loop statement.

Strictly follow the format: "<n> is <a/NOT a> Germain Prime Number."

Samples
input: 5
output: 5 is a Germain Prime Number.
input: 7
output: 7 is NOT a Germain Prime Number.
input: 29
output: 29 is a Germain Prime Number.
input: 59
output: 59 is NOT a Germain Prime Number.