Create MACHINE PROBLEM #8.3 - sArCaStIc cAsE.c
Sarcasm on the Internet is difficult to convey, but one way to convey it is by changing the letter case alternately. "yOu ArE sO kInD" is an example of sarcastically complimenting another of their kindness.

Given a string, convert it into sArCaSm. At most, there'll be 70 characters in the string. The words in the string will be comprised of the letters of the English alphabet.

You are not allowed to use any function expect for scanf, printf, or your own function.

Always start words with lowercase and then alternate the case. For instance, the string "Hello world" should be converted into: "hElLo wOrLd".

Accepting spaces in the string
Since scanf("%s", ...); will stop reading input on seeing a space, and thus only reading a single word, you should use the following format: scanf("%69[^\n]", ...);. This format means scanf will accept 69 characters except the \n character. When the newline character is seen, scanf will end.

Samples
input: Hello world
output: hElLo wOrLd
input: wow
output: wOw
input: You are so brilliant
output: yOu aRe sO bRiLlIaNt